+++ CHANGELOG.md
@@ -5,5 +5,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v1.0.1] - 2016-06-21
+- Update docs, docblocks, and fix homepage link
+- Add per repository cache lifetime/driver support (Close #10)
+
 ## v1.0.0 - 2016-06-18
 - Commit first draft
+
+[v1.0.1]: https://github.com/rinvex/repository/compare/v1.0.0...v1.0.1